ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
scikit-learn
google-generativeai
